refactor leaderboard_embed  & interactable_leaderboard_embed
